Inversion exit rule
