+++ src/app/exercise/exercises/ChordInKeyExercise/grammar/chord-progression-rules.ts
@@ -2,9 +2,11 @@ import { RomanNumeralChordSymbol } from '../../../utility';
 import { Rule, noRepeatsRule } from '../../../utility/grammer';
 import { dominantResolutionRule } from './rules/dominant-resolution-rule';
 import { inversionEnterRule } from './rules/inversion-enter-rule';
+import { inversionExitRule } from './rules/inversion-exit-rule';
 
 export const chordProgressionRules: Rule<RomanNumeralChordSymbol>[] = [
   noRepeatsRule,
   dominantResolutionRule,
   inversionEnterRule,
+  inversionExitRule,
 ];
