+++ src/app/exercise/exercises/ChordInKeyExercise/grammar/utilities/is-step-or-unison.ts
@@ -1,0 +1,13 @@
+import { ScaleDegree, scaleDegreeToChromaticDegree } from '../../../../utility';
+
+export function isStepOrUnison(
+  scaleDegree1: ScaleDegree,
+  scaleDegree2: ScaleDegree,
+) {
+  return (
+    Math.abs(
+      scaleDegreeToChromaticDegree[scaleDegree1] -
+        scaleDegreeToChromaticDegree[scaleDegree2],
+    ) <= 2
+  );
+}
